2080_LLM/Refactored_Psy/NoAug_Criteria_Evidence/pyproject.toml
[tool.poetry]
name = "psy_agents_noaug"
version = "0.1.0"
description = "NO-AUG baseline for Criteria and Evidence extraction from clinical text"
authors = ["Research Team"]
readme = "README.md"
packages = [{include = "psy_agents_noaug", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
torch = "^2.0.0"
transformers = "^4.30.0"
hydra-core = "^1.3.0"
mlflow = "^2.5.0"
optuna = "^3.2.0"
pandas = "^2.0.0"
numpy = "^1.24.0"
scikit-learn = "^1.3.0"
datasets = "^2.13.0"
nlpaug = "^1.1.11"  # For comparison only - NOT used in this repo
peft = "^0.10.0"
pyyaml = "^6.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.1.0"
black = "^23.7.0"
isort = "^5.12.0"
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.11.1"
pre-commit = "^3.3.0"
bandit = "^1.7.5"

[tool.poetry.scripts]
psy-noaug = "psy_agents_noaug.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3

[tool.ruff]
line-length = 88
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]  # Line too long (handled by black)
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

[tool.coverage.run]
source = ["src/psy_agents_noaug"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/conftest.py",
]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "def __str__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
precision = 2
show_missing = true

[tool.coverage.html]
directory = "htmlcov"

[tool.bandit]
exclude_dirs = ["tests", ".venv", "venv"]
skips = ["B101"]  # assert_used - we use asserts in tests
